fix: 상세페이지로 안넘어가지는 문제 해결 + 빌드 에러 수정
